requirements.txt
fastapi>=0.103.0,<0.104.0
pydantic==2.4.2
pytest>=7.4.0,<7.5.0
httpx>=0.24.0,<0.25.0
pydantic-settings==2.0.3
uvicorn>=0.23.0,<0.24.0
sqlalchemy>=2.0.0,<3.0.0
python-jose>=3.3.0,<3.4.0
passlib>=1.7.4,<1.8.0
python-multipart>=0.0.5,<0.0.6
email-validator>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
psycopg2-binary>=2.9.6,<2.10.0
alembic>=1.12.0,<1.13.0